Fix 10b_preview parallel execution and schema alignment
Enable preview mode to skip stages already completed by earlier ops pipeline
steps, preventing parallel conflicts and ensuring 10b_preview tests only the
gold upsert SELECT paths as intended.

Changes:
- runner.py: Skip PG_FUNCTIONS_INSTALL, BRONZE_*, SILVER_NORMALISE, DBT_BUILD,
  and STACKSYNC_SYNC in preview mode to avoid concurrent DDL and redundant work
- 10b_preview.ps1: Add --skip-validation flag since 07_dbt already ran
- probe_schemas.py: Update PK column names to match icalps_* naming convention
- 06_silver.ps1: Convert to validation-only mode (row count checks)
- pyproject.toml: Switch dbt adapter from postgres to duckdb
- sources.yml: Add database: pg_hubspot for attached PostgreSQL
- stg_hubspot_*.sql: Cast legacy IDs to BIGINT to avoid overflow
- stg_bronze_communication.sql: Fix column case to match actual schema
- dbt tests: Add pg_hubspot schema prefix and fix column references

+++ dbt/models/staging/stg_hubspot_companies.sql
@@ -3,7 +3,7 @@
 
 select
     id as hubspot_company_id,
-    cast(icalps_company_id as integer) as legacy_company_id,
+    cast(icalps_company_id as bigint) as legacy_company_id,
     stacksync_record_id_9vpp8v as hubspot_company_record_id,
     name as company_name,
     domain as company_domain,
